vllm/requirements.txt
transformers==4.45.1
peft==0.13.0
datasets==3.0.1
trl==0.11.1
bitsandbytes==0.42.0
pillow
evaluate==0.4.3
accelerate==0.34.2
qwen_vl_utils
tensorboardX
